dette d'echelle v6 : plafonds declares, cap de cardinalite a overshoot borne, budget partiel de tampons
checkpoint accepte par l'auditeur (sept jets en vol absorbes) : caps.hpp
(plafonds structurels — arbre radix 2^30-1 refuse avant construction,
candidats bruts u32 remonte a l'emission ; fits_budget a produits
controles) ; cap de cardinalite a overshoot borne 4096 x T (throttle aux
trois sites d'emission, arret remonte a la boucle d'ancres, somme exacte
et reserve unique avant la fusion globale, shards liberes des leur
fusion — le pic double de la fusion-copie est mort) ; plafonds de vague
prospectifs pre-insertion dans wave/next/out (vague initiale comprise,
pics observes) ; budget PARTIEL de tampons nommes = proxy de payload
logique (tri x2, prefiltre/census conservatif S<=C avec BallData x2,
evenements du fold x (inflight+2)), cap effectif derive par helper
partage et signe par la CLI (memory_budget_scope=..., porte UNIQUE
code+ligne via run_expect etendu — PASS_REGULAR ignore le code) ;
familles : quatre gardes 200/400*n en long long (UB des n=5,37M — les
familles de pente), quota (n+7)/8 elargi, --n borne a 2^30-1 des la CLI ;
porte --caps-refus : temoin + fits_budget (egalite / -1 / overflows), 5
etages causaux (emission, budget->emission, tri a tri_need-1, census a
tri_need, fold a smax=2), caps de test wave/alive, overshoot observable
borne, zero callback/phase sur chaque refus, objet identique sous budget
large ; mutants caps-drop-emission et caps-late-wave-check (moment de
garde) tues. suite hors echelle 77+2/77+2, docs 242, diff propre.
v5 NON touchee (note auditeur : requalification separee requise).

+++ morsehgp3D_v6/src/cloud/families.hpp
@@ -140,7 +140,7 @@ inline std::vector<P3> uniform_cloud(int n, int coord, long long seed) {
   std::uniform_int_distribution<int> pick(0, coord - 1);
   std::vector<P3> pts;
   std::set<long long> keys;
-  for (int guard = 0; (int)pts.size() < n && guard < 200 * n; ++guard) {
+  for (long long guard = 0; (int)pts.size() < n && guard < 200LL * n; ++guard) {
     P3 q{};
     q.x = (i64)pick(rng);
     q.y = (i64)pick(rng);
@@ -163,8 +163,8 @@ inline std::vector<P3> eight_clusters_cloud(int n, int coord, long long seed) {
   std::vector<P3> pts;
   std::set<long long> keys;
   int placed[8] = {0, 0, 0, 0, 0, 0, 0, 0};
-  const int quota = (n + 7) / 8;
-  for (int guard = 0; (int)pts.size() < n && guard < 400 * n; ++guard) {
+  const int quota = (int)(((long long)n + 7) / 8);
+  for (long long guard = 0; (int)pts.size() < n && guard < 400LL * n; ++guard) {
     int c = -1;  // tourniquet strict : amas equilibres a un point pres
     for (int t = 0; t < 8; ++t) {
       const int cand = (guard + t) % 8;
@@ -219,7 +219,7 @@ inline std::vector<P3> terrain_cloud(int n, int coord, long long seed, int canop
   std::uniform_int_distribution<int> canopy_lift(1, lift_max);
   std::vector<P3> pts;
   std::set<long long> keys;
-  for (int guard = 0; (int)pts.size() < n && guard < 200 * n; ++guard) {
+  for (long long guard = 0; (int)pts.size() < n && guard < 200LL * n; ++guard) {
     const long long x = ground(rng);
     const long long y = ground(rng);
     long long h = 0;
@@ -338,7 +338,7 @@ inline std::vector<P3> scanline_cloud(int n, int coord, long long seed, bool ove
   const int pitch = 8;  // anisotropie 4:1 entre l'inter-ligne et le pas
   std::vector<P3> pts;
   std::set<long long> keys;
-  scanline_pass(&rng, field, overlap_multiecho ? (n * 3) / 5 : n, coord, 0, 1, step_along, pitch,
+  scanline_pass(&rng, field, overlap_multiecho ? (int)(((long long)n * 3) / 5) : n, coord, 0, 1, step_along, pitch,
                 overlap_multiecho, &pts, &keys);
   if (overlap_multiecho)
     scanline_pass(&rng, field, n, coord, step_along / 2 + 1, pitch / 3 + 1, step_along, pitch, true,
@@ -412,7 +412,7 @@ inline std::vector<P3> terrain_stationnaire_cloud(int n, int coord, long long se
   std::uniform_int_distribution<int> canopy_lift(1, 55);  // = [1, max(2, c0/8)]
   std::vector<P3> pts;
   std::set<long long> keys;
-  for (int guard = 0; (int)pts.size() < n && guard < 200 * n; ++guard) {
+  for (long long guard = 0; (int)pts.size() < n && guard < 200LL * n; ++guard) {
     const long long x = ground(rng);
     const long long y = ground(rng);
     long long h = 0;
